Clarity in CM
Cleaned up the code for Consensus Mechanism and the Tests, simplified
FillNA function and Participation variables, added the Figures to the
Tests for additional clarity, Figures now appropriately reference the
FillNA'ed version of their inputs, changed the colours for Figures for
greater visibility. Ran and saved tests so there would be reference
figures.

Added somes notes to docs, as well.