Contract Data Structure
Substantially improved Contracts.r: The contract object probably all of
the fields it would need for the reference implementation, which
includes fields for trading and accounting.

I still need to modify Trading so that it works correctly with these new
changes. Then I will need to modify Governance to tie all of the pieces
together and produce a block validation function that will mostly
complete version 0.1.